ArtistService.getArtist now accepts a long

+++ app/src/main/java/com/voxxel/api/ArtistService.java
@@ -11,5 +11,5 @@ public interface ArtistService {
 
     @GET("/api/v1/artists/{id}")
     @Headers("Content-Type:application/json")
-    ArtistModel getArtist(@Path("id") int id);
+    ArtistModel getArtist(@Path("id") Long id);
 }
